Edited the pseudo code
This part:
o	If the total profit is greater than or equal to $10,000, then the program displays the message: “You did well this period! Keep up the great work!”
o	If the total profit is less than $10,000, the program displays the message: “We didn’t reach our goal for this period. More work is needed.”
•	The program continues working until the user exits by entering 0.